Refactor unit testing
