+++ src/poly.h
@@ -33,11 +33,7 @@ typedef struct {
 } poly;
 
 void eval_poly(fr_t *out, const poly *p, const fr_t *x);
-C_KZG_RET poly_long_div(poly *out, const poly *dividend, const poly *divisor);
-C_KZG_RET poly_mul_direct(poly *out, const poly *a, const poly *b);
-C_KZG_RET poly_mul_fft(poly *out, const poly *a, const poly *b, FFTSettings *fs);
 C_KZG_RET poly_inverse(poly *out, poly *b);
-C_KZG_RET poly_fast_div(poly *out, const poly *dividend, const poly *divisor);
 C_KZG_RET poly_mul(poly *out, const poly *a, const poly *b);
 C_KZG_RET poly_mul_(poly *out, const poly *a, const poly *b, FFTSettings *fs);
 C_KZG_RET new_poly_div(poly *out, const poly *dividend, const poly *divisor);
